audit-tumble-engine-bitexact-literals: s3 repair — fix tautological assertions, citations, and enumeration command
Blocker 1: 17 tautological assertions (assert_eq!(A, A)) removed. Module-level
consts in other files (math.rs ATAN_P0-P3, finalize.rs SPECULATIVE_DISTANCE/
POSITION_SLEEP_FACTOR/SAFETY_FACTOR, integrate.rs MAX_ROTATION, recycle.rs
RECYCLE_ANGULAR_DISTANCE) now have c_parity modules appended to their own files,
reading super::THE_CONST directly. Function-local lets and inline literals
(manifold.rs:212/648/1003/1309/1402/1929, math.rs:138/141/350/351) are marked
"not assertable" in the table with values verified by hand against C.

Blocker 2: Comment no longer claims exact k/2^n literals are "excluded from
assertions" while asserting POSITION_SLEEP_FACTOR (0.5) and FLT_EPSILON (2^-23).
One rule: assert exactly when the subject is a readable source item; exactness
is an annotation, never an exclusion criterion.

Defect 3: B3_MAX_ROTATION citation fixed from constants.h:75 (a comment about
B3_MESH_REST_OFFSET) to constants.h:70 (the actual #define).

Defect 4: OVERLAP_SLOP reachability citations fixed: b3OverlapCapsule
capsule.c:84 (was :94), b3OverlapHull hull.c:2439 (was :2451), b3OverlapSphere
sphere.c:47 (was :57). All citations now include the symbol name beside the
line number.

Note 6: Enumeration command replaced with sed '/#\[cfg(test)\]/,$d' which does
not depend on the test-module marker being present (the old grep|head -1 shape
silently became head -n -1 for files without #[cfg(test)]).

cargo test: 71 pass / 0 fail (was 80 with 25 c_parity tautologies; now 16 real
assertions across 5 files + 55 non-c_parity = 71).
